backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
mysqlclient==2.2.7